Refs #13040 Additional fix for gcc warning

+++ Code/Mantid/Framework/DataHandling/src/LoadBBY.cpp
@@ -190,7 +190,7 @@ void LoadBBY::exec() {
           mask[s0 + (size_t)y] = false;
       }
       else { // if (offset < 0)
-        for (int y = (HISTO_BINS_Y + static_cast<size_t>(offset)); y != HISTO_BINS_Y; y++)
+        for (size_t y = (HISTO_BINS_Y + static_cast<size_t>(offset)); y != HISTO_BINS_Y; y++)
           mask[s0 + y] = false;
       }
     }
